Example #3: Add DBUS example
In this example, dbussession and dbusclient packages have been created to communicate each other over
system bus DBUS. New_dbus_policy.conf file is needed to add a policy to use system bus as a
"root" user. This is done with post build script.

As an extra, dbus library is selected with menuconfig and saved under configs folder.

Signed-off-by: Azad Karatas <azadkaratas@gmail.com>